fix: address endpoint discovery review threads
Split MCP schema request models, align the CommandKind contract docs, and keep endpoint discovery request/help contracts current.

Resolves review thread PRRT_kwDORS2O8s6DW1vr

Resolves review thread PRRT_kwDORS2O8s6DW1vu

Resolves review thread PRRT_kwDORS2O8s6DW1vy

Resolves review thread PRRT_kwDORS2O8s6DW1v1

Resolves review thread PRRT_kwDORS2O8s6DW1v4

Resolves review thread PRRT_kwDORS2O8s6DW1wD

Resolves review thread PRRT_kwDORS2O8s6DW1wE
